app/requirements.txt
flask
python-dotenv
azure-storage-blob
sqlalchemy